+++ Tests/Finbuckle.SharedDatabase.TestApplication/Finbuckle.SharedDatabase.TestApplication.Application/DependencyInjection.cs
@@ -4,6 +4,7 @@ using Finbuckle.SharedDatabase.TestApplication.Application.Implementation;
 using Finbuckle.SharedDatabase.TestApplication.Application.Interfaces;
 using FluentValidation;
 using Intent.RoslynWeaver.Attributes;
+using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
@@ -17,6 +18,7 @@ namespace Finbuckle.SharedDatabase.TestApplication.Application
         {
             services.AddValidatorsFromAssembly(Assembly.GetExecutingAssembly());
             services.AddAutoMapper(Assembly.GetExecutingAssembly());
+            services.AddMediatR(Assembly.GetExecutingAssembly());
             services.AddTransient<IValidationService, ValidationService>();
             services.AddTransient<IUsersService, UsersService>();
             return services;