+++ Tests/Finbuckle.SharedDatabase.TestApplication/Finbuckle.SharedDatabase.TestApplication.Domain/Entities/User.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using Finbuckle.SharedDatabase.TestApplication.Domain.Common;
 using Intent.RoslynWeaver.Attributes;
 
 [assembly: DefaultIntentManaged(Mode.Fully)]
@@ -10,7 +11,7 @@ namespace Finbuckle.SharedDatabase.TestApplication.Domain.Entities
     [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
     [DefaultIntentManaged(Mode.Fully, Targets = Targets.Properties)]
     [DefaultIntentManaged(Mode.Fully, Targets = Targets.Methods, Body = Mode.Ignore, AccessModifiers = AccessModifiers.Public)]
-    public class User
+    public class User : IHasDomainEvent
     {
         public Guid Id { get; set; }
 
@@ -19,5 +20,7 @@ namespace Finbuckle.SharedDatabase.TestApplication.Domain.Entities
         public string Username { get; set; }
 
         public virtual ICollection<Role> Roles { get; set; } = new List<Role>();
+
+        public List<DomainEvent> DomainEvents { get; set; } = new List<DomainEvent>();
     }
 }